Users
update the methods around adding and deleting user. In particular,  the absence of a clerkID for users, will indicate that the user has been added by the admin.  When the user logs in via  Clerk the user is prompted to check/edit their information, and then a clerkID will be attached to that user.