Update EditCourseInfo.tsx // Bugs and Issues
Refactored a useEffect used to detect changes in the course by use of deep comparison. Now it sorts classes of both incoming and edited course, effectively making the order of the courses class not important. This is because the order of the classes from the database are not guaranteed.  I will change this in the future, but for now I'm going to handle the possibility of this situation. Tangentially, I could make trigger the detected change warning pop up whenever any changes happen at all, however if the user changes the state of the course back to It's original state, it would not detect that and render a warning anyway.