Update editCourseInfo.tsx // Bugs and Issues
Tried to change the order of methods in attempt to change the fault, but  I didn't notice any changes.

Additionally, The issues seems to be when the class's scheduleStart value changes, the parent component wants to reorder the classes immediately. I think this is causing problems because the user is still focused on the  class  in the original index.