Updated EditcourseInfo.tsx // Bugs and Issues
The ordering bug has been resolved by only ordering a course's classes up initial load. Maybe at a later time I will have logic that will update the order of the class more dynamically, but for now this will work.

I was then able to recreate the fault of a pervious bug. The original bug deleted classes when classes were in a modified order and the scheduleStart values were changed.