Bug Resolved
The bug was due to using the index and the containing div's key. which would have kept the ClassCard from re-rendering with the right class data.
However, if the user tries to adds multiple new classes, which would have the classId of 0, it will log a warning to the console for multiple elements with the same key.