WIP Saving Course
Work in progress commit, when saving, since the incoming courses is a prop, thus is will never change and so it will never reset the change detection warning. Im going to have to make a change in how this component gets its course data, moving away from using the prop.